requirements.txt
prefect
minio
pandas
pyarrow
faker
streamlit
plotly
python-dotenv
fastapi
uvicorn[standard]
pymongo
motor
httpx